* subr.c (copy_file): Declare local var n to be an int, not a size_t.  size_t is unsigned, and the return values of read and write are definitely not unsigned.
* cvs.h [HAVE_IO_H]: #include <io.h>.
[HAVE_DIRECT_H]: #include <direct.h>.

CVS-ID: src/cvs.h 1.23
CVS-ID: src/subr.c 1.23

+++ src/subr.c
@@ -77,12 +77,12 @@ copy_file (from, to)
     if (sb.st_size > 0)
     {
 	char buf[BUFSIZ];
-	size_t n;
+	int n;
 
 	for (;;) 
 	{
 	    n = read (fdin, buf, sizeof(buf));
-	    if (n == (size_t)-1)
+	    if (n == -1)
 	    {
 #ifdef EINTR
 		if (errno == EINTR)
